do not treat 304 as good status code by default
HTTP servers MUST NOT include any content when sending a 304. When treating
304 as success, the response body is empty and transforming the response
leads to an empty string. This then fails in `assureResponse` when
`response.data` is assigned to the variable of type `T` (which is usually
not String, but the expected body type like `List<dynamic>`) with an
exception: type 'String' is not a subtype of type 'List<dynamic>'

By treating 304 as a failure, the request throws a DioError that can
easily be checked for the statusCode and then the application can handle
the 304 however it wants to (usually by reusing the data from the cache).