Add `native_diox_adapter` plugin (#49)
This is the code from my
[`native_dio_client`](https://pub.dev/packages/native_dio_client)
package, but adapted to `diox`.

I've changed the name from `client` to `adapter` to better reflect the
dio terms.
The code unfortunately isn't really tested, but it's tested in
production.
[`native_dio_client`](https://pub.dev/packages/native_dio_client) has a
popularity of about 80%.

The plugin works like this:

The main logic is within the `ConversionLayerAdapter` class. That class
basically converts all diox requests to requests compatible with the
[`http`](https://pub.dev/packages/http) lib. Both `cupertino_http` and
`cronet_http` provide `http` compatible clients. This keeps the
maintenance effort low as we don't need to implement individual adapter
for both libraries.

Because of the way `cupertino_http` and `cronet_http` work, the plugin
has a dependency on Flutter and `http`.

Signed-off-by: Jonas Uekötter <ueman@users.noreply.github.com>
Co-authored-by: Alex Li <github@alexv525.com>