Fix "unsupported operation" error on web platform (#1698)
For the web platform, an inline `bool.fromEnvironment` translates to
`core.bool.fromEnvironment` instead of correctly being replaced by the
constant value found in the environment at build time.

See https://github.com/flutter/flutter/issues/51186.

### New Pull Request Checklist

- [x] I have read the
[Documentation](https://pub.dev/documentation/dio/latest/)
- [x] I have searched for a similar pull request in the
[project](https://github.com/cfug/dio/pulls) and found none
- [x] I have updated this branch with the latest `main` branch to avoid
conflicts (via merge from master or rebase)
- [ ] I have added the required tests to prove the fix/feature I'm
adding
- [x] I have updated the documentation (if necessary)
- [x] I have run the tests without failures
- [x] I have updated the `CHANGELOG.md` in the corresponding package (if
necessary)

### Additional context and info (if any)

This fix is unfortunately not something we can test for as far as I can
tell. I've tried numerous methods of automating the detection of the
issue (and thus the fix), but it seems to only appear when running the
Flutter example in a browser.